Refactor: Make it harder to have symbol conflict
With VERIFY() being a macro it was easy to run into conflicts with
other libraries which might want to implement similar functionality.

Now Verify() is a function, and follows function naming convention.
This will help preventing any "leaks" of API to global namespace,
and should make it harder to conflict in name style with macros used
by other libraries.

The downside of this is that it is not possible to convert expression
to string automatically and message is to be provided explicitly.
This is similar to Python's assert statement.

Similarly, make Unreachable() to follow function naming, which should
help avoiding conflicts with macro names from other libraries.