Add Python tools to visalize constellation and FFT of WAV
The constellation plotter reads a stereo WAV file as complex numbers
and visualized them as constellation. It is possible to control the
number of points in the constellation, its scale, and the number of
new samples read from the file on every redraw.

The constellation plotter is using intensity grading.

The FFT reads a stereo WAV file as complex numbers, calculates FFT
and visualizes it.