Integration with FFT libraries
This commit lays down an initial foundation for a common API for
interfacing with external FFT libraries. It is currently implemented
for PFFFT library. Supports complex-to-complex and real-to-complex
transformation.

The FFT libraries are not intended to be shipped with this library
so that it can stay header-only, and the user might decide to use a
different FFT library.